+++ cibc/src/Packages/MatlabInterface/Dataflow/Modules/DataIO/MatlabNrrdsReader.cc
@@ -94,7 +94,7 @@ class MatlabNrrdsReader : public Module
     //   retrieves the current filename and the matrix selected in this
     //   file and returns an object containing the full matrix
     
-    matlabarray readmatlabarray(long p);
+    matlabarray readmatlabarray(int p);
 
     // displayerror()
     //   Relay an error during the matrixselection process directly to
@@ -177,7 +177,7 @@ void MatlabNrrdsReader::execute()
 
   try
   {
-    for (long p=0;p<NUMPORTS;p++)
+    for (int p=0;p<NUMPORTS;p++)
     {
 
       // Find the output port the scheduler has created 
@@ -284,7 +284,7 @@ void MatlabNrrdsReader::tcl_command(GuiArgs& args, void* userdata)
 }
 
 
-matlabarray MatlabNrrdsReader::readmatlabarray(long p)
+matlabarray MatlabNrrdsReader::readmatlabarray(int p)
 {
   matlabarray marray;
   std::string filename = guifilename_.get();
@@ -361,7 +361,7 @@ void MatlabNrrdsReader::indexmatlabfile(bool postmsg)
   std::vector<std::string> matrixnamelist(NUMPORTS);
   bool foundmatrixname[NUMPORTS];
 
-  for (long p=0;p<NUMPORTS;p++)
+  for (int p=0;p<NUMPORTS;p++)
   {
     // TCL Dependent code
     std::ostringstream oss;
@@ -383,15 +383,15 @@ void MatlabNrrdsReader::indexmatlabfile(bool postmsg)
     
     // all matlab data is stored in a matlabarray object
     matlabarray ma;
-    long cindex = 0;		// compatibility index, which matlab array fits the SCIRun Nrrd best? 
-    long maxindex = 0;		// highest index found so far
+    int cindex = 0;		// compatibility index, which matlab array fits the SCIRun Nrrd best? 
+    int maxindex = 0;		// highest index found so far
             
     // Scan the file and see which matrices are compatible
     // Only those will be shown (you cannot select incompatible matrices).
             
     std::string infotext;
     
-    for (long p=0;p<mfile.getnummatlabarrays();p++)
+    for (int p=0;p<mfile.getnummatlabarrays();p++)
     {
       ma = mfile.getmatlabarrayinfo(p); // do not load all the data fields
       if ((cindex = translate.sciNrrdDataCompatible(ma,infotext)))
@@ -404,7 +404,7 @@ void MatlabNrrdsReader::indexmatlabfile(bool postmsg)
         
         matrixinfotexts += std::string("{" + infotext + "} ");
         matrixnames += std::string("{" + ma.getname() + "} "); 
-        for (long q=0;q<NUMPORTS;q++)
+        for (int q=0;q<NUMPORTS;q++)
         {
           if (ma.getname() == matrixnamelist[q]) foundmatrixname[q] = true;
         }
@@ -420,7 +420,7 @@ void MatlabNrrdsReader::indexmatlabfile(bool postmsg)
     // cannot be found or if no matrixname has been specified
     
     matrixname = "";
-    for (long p=0;p<NUMPORTS;p++)
+    for (int p=0;p<NUMPORTS;p++)
     {
       if (foundmatrixname[p] == false) 
       {
